deprecated some ETL pipelines that are no longer in use
