+++ src/main/java/edu/cuanschutz/ccp/tm_provider/etl/deprecated/BigQueryExportPipeline.java
@@ -1,4 +1,4 @@
-package edu.cuanschutz.ccp.tm_provider.etl;
+package edu.cuanschutz.ccp.tm_provider.etl.deprecated;
 
 import java.util.ArrayList;
 import java.util.Arrays;
@@ -26,10 +26,11 @@ import org.apache.beam.sdk.values.PCollectionTuple;
 
 import com.google.datastore.v1.Entity;
 
+import edu.cuanschutz.ccp.tm_provider.etl.EtlFailureData;
+import edu.cuanschutz.ccp.tm_provider.etl.PipelineMain;
 import edu.cuanschutz.ccp.tm_provider.etl.fn.BigQueryExportFileBuilderFn;
 import edu.cuanschutz.ccp.tm_provider.etl.fn.DocumentDownloadFn;
 import edu.cuanschutz.ccp.tm_provider.etl.fn.EtlFailureToEntityFn;
-import edu.cuanschutz.ccp.tm_provider.etl.fn.PubAnnotationExportFileBuilderFn;
 import edu.cuanschutz.ccp.tm_provider.etl.util.DatastoreProcessingStatusUtil;
 import edu.cuanschutz.ccp.tm_provider.etl.util.DatastoreProcessingStatusUtil.OverwriteOutput;
 import edu.cuanschutz.ccp.tm_provider.etl.util.DocumentCriteria;
@@ -40,13 +41,13 @@ import edu.cuanschutz.ccp.tm_provider.etl.util.ProcessingStatusFlag;
 import edu.cuanschutz.ccp.tm_provider.etl.util.Version;
 
 /**
- * This Apache Beam pipeline exports annotations in the PubAnnotation format.
+ * NOTE: This pipeline is not actively used.
  */
-public class PubAnnotationExportPipeline {
+public class BigQueryExportPipeline {
 
-	private final static Logger LOGGER = Logger.getLogger(PubAnnotationExportPipeline.class.getName());
+	private final static Logger LOGGER = Logger.getLogger(BigQueryExportPipeline.class.getName());
 
-	private static final PipelineKey PIPELINE_KEY = PipelineKey.PUBANNOTATION_EXPORT;
+	private static final PipelineKey PIPELINE_KEY = PipelineKey.BIGQUERY_EXPORT;
 
 	public interface Options extends DataflowPipelineOptions {
 		@Description("Location of the output bucket")
@@ -63,14 +64,13 @@ public class PubAnnotationExportPipeline {
 		OverwriteOutput getOverwrite();
 
 		void setOverwrite(OverwriteOutput value);
-
 	}
 
 	public static void main(String[] args) {
 		String pipelineVersion = Version.getProjectVersion();
 		com.google.cloud.Timestamp timestamp = com.google.cloud.Timestamp.now();
 		Options options = PipelineOptionsFactory.fromArgs(args).withValidation().as(Options.class);
-		LOGGER.log(Level.INFO, String.format("Running PubAnnontation export pipeline"));
+		LOGGER.log(Level.INFO, String.format("Running BigQuery export pipeline"));
 
 		Pipeline p = Pipeline.create(options);
 		PCollection<String> documentIds = getDocumentIdsToProcess(p, options.getCollection(), options.getOverwrite());
@@ -94,37 +94,58 @@ public class PubAnnotationExportPipeline {
 		nonredundantFailureEntities.apply("annot_extract_failure_entity->datastore",
 				DatastoreIO.v1().write().withProjectId(options.getProject()));
 
-		DocumentCriteria outputDocCriteria = new DocumentCriteria(DocumentType.PUBANNOTATION, DocumentFormat.JSON,
+		DocumentCriteria outputDocCriteria = new DocumentCriteria(DocumentType.BIGQUERY, DocumentFormat.BIGQUERY,
 				PIPELINE_KEY, pipelineVersion);
 
-		PCollectionTuple pubAnnotExports = PubAnnotationExportFileBuilderFn.processNoTrackDocIds(docIdToAnnotations,
+		PCollectionTuple bigqueryExports = BigQueryExportFileBuilderFn.processNoTrackDocIds(docIdToAnnotations,
 				outputDocCriteria, timestamp);
 
-		PCollection<String> docIdToJson = pubAnnotExports.get(PubAnnotationExportFileBuilderFn.JSON_OUTPUT_TAG);
-		PCollection<EtlFailureData> pubAnnotExportFailures = pubAnnotExports
+		PCollection<String> docIdToBigQuery_annotationTable = bigqueryExports
+				.get(BigQueryExportFileBuilderFn.ANNOTATION_TABLE_OUTPUT_TAG_NO_TRACK_DOCID);
+		PCollection<String> docIdToBigQuery_inSectionTable = bigqueryExports
+				.get(BigQueryExportFileBuilderFn.IN_SECTION_TABLE_OUTPUT_TAG_NO_TRACK_DOCID);
+		PCollection<String> docIdToBigQuery_inParagraphTable = bigqueryExports
+				.get(BigQueryExportFileBuilderFn.IN_PARAGRAPH_TABLE_OUTPUT_TAG_NO_TRACK_DOCID);
+		PCollection<String> docIdToBigQuery_inSentenceTable = bigqueryExports
+				.get(BigQueryExportFileBuilderFn.IN_SENTENCE_TABLE_OUTPUT_TAG_NO_TRACK_DOCID);
+		PCollection<String> docIdToBigQuery_inConceptTable = bigqueryExports
+				.get(BigQueryExportFileBuilderFn.IN_CONCEPT_TABLE_OUTPUT_TAG_NO_TRACK_DOCID);
+		PCollection<String> docIdToBigQuery_relationTable = bigqueryExports
+				.get(BigQueryExportFileBuilderFn.RELATION_TABLE_OUTPUT_TAG_NO_TRACK_DOCID);
+		PCollection<EtlFailureData> bigqueryExportFailures = bigqueryExports
 				.get(BigQueryExportFileBuilderFn.FAILURE_TAG);
 
 		/*
 		 * store the failures for this pipeline in Cloud Datastore - deduplication is
 		 * necessary to avoid Datastore non-transactional commit errors
 		 */
-		failureEntities = pubAnnotExportFailures.apply("pubannot_export_failures->datastore",
+		failureEntities = bigqueryExportFailures.apply("bq_export_failures->datastore",
 				ParDo.of(new EtlFailureToEntityFn()));
 		nonredundantFailureEntities = PipelineMain.deduplicateEntitiesByKey(failureEntities);
-		nonredundantFailureEntities.apply("pubannot_export_failure_entity->datastore",
+		nonredundantFailureEntities.apply("bq_export_failure_entity->datastore",
 				DatastoreIO.v1().write().withProjectId(options.getProject()));
 
 		// write the output
-		docIdToJson.apply("write pubannot json file",
-				TextIO.write().to(options.getOutputBucket() + "/pubannotation.").withSuffix(".json"));
+		docIdToBigQuery_annotationTable.apply("write annotation table",
+				TextIO.write().to(options.getOutputBucket() + "/annotation.").withSuffix(".tsv"));
+		docIdToBigQuery_inSectionTable.apply("write annotation table",
+				TextIO.write().to(options.getOutputBucket() + "/in-section.").withSuffix(".tsv"));
+		docIdToBigQuery_inParagraphTable.apply("write annotation table",
+				TextIO.write().to(options.getOutputBucket() + "/in-paragraph.").withSuffix(".tsv"));
+		docIdToBigQuery_inSentenceTable.apply("write annotation table",
+				TextIO.write().to(options.getOutputBucket() + "/in-sentence.").withSuffix(".tsv"));
+		docIdToBigQuery_inConceptTable.apply("write annotation table",
+				TextIO.write().to(options.getOutputBucket() + "/in-concept.").withSuffix(".tsv"));
+		docIdToBigQuery_relationTable.apply("write annotation table",
+				TextIO.write().to(options.getOutputBucket() + "/relation.").withSuffix(".tsv"));
 
 		PCollectionList<EtlFailureData> failureList = PCollectionList.of(annotationRetrievalFailures)
-				.and(pubAnnotExportFailures);
+				.and(bigqueryExportFailures);
 		PCollection<EtlFailureData> mergedFailures = failureList.apply(Flatten.<EtlFailureData>pCollections());
 
 		// update the status for documents that were successfully processed
 		PCollection<KV<String, String>> successStatus = DatastoreProcessingStatusUtil.getSuccessStatus(documentIds,
-				mergedFailures, ProcessingStatusFlag.PUBANNOTATION_FILE_EXPORT_DONE);
+				mergedFailures, ProcessingStatusFlag.BIGQUERY_LOAD_FILE_EXPORT_DONE);
 		List<PCollection<KV<String, String>>> statusList = new ArrayList<PCollection<KV<String, String>>>();
 		statusList.add(successStatus);
 		DatastoreProcessingStatusUtil.performStatusUpdatesInBatch(statusList);
@@ -136,10 +157,12 @@ public class PubAnnotationExportPipeline {
 	private static List<DocumentCriteria> populateDocumentCriteria(String pipelineVersion) {
 		List<DocumentCriteria> documentCriteria = Arrays.asList(
 				new DocumentCriteria(DocumentType.TEXT, DocumentFormat.TEXT, PipelineKey.BIOC_TO_TEXT, pipelineVersion),
-//				new DocumentCriteria(DocumentType.SECTIONS, DocumentFormat.BIONLP, PipelineKey.BIOC_TO_TEXT,
-//						pipelineVersion),
-//				new DocumentCriteria(DocumentType.DEPENDENCY_PARSE, DocumentFormat.CONLLU, PipelineKey.DEPENDENCY_PARSE,
-//						pipelineVersion),
+				new DocumentCriteria(DocumentType.SENTENCE, DocumentFormat.BIONLP, PipelineKey.SENTENCE_SEGMENTATION,
+						pipelineVersion),
+				new DocumentCriteria(DocumentType.SECTIONS, DocumentFormat.BIONLP, PipelineKey.BIOC_TO_TEXT,
+						pipelineVersion),
+				new DocumentCriteria(DocumentType.DEPENDENCY_PARSE, DocumentFormat.CONLLU, PipelineKey.DEPENDENCY_PARSE,
+						pipelineVersion),
 				new DocumentCriteria(DocumentType.CONCEPT_CHEBI, DocumentFormat.BIONLP, PipelineKey.OGER,
 						pipelineVersion),
 				new DocumentCriteria(DocumentType.CONCEPT_CL, DocumentFormat.BIONLP, PipelineKey.OGER, pipelineVersion),
@@ -169,12 +192,17 @@ public class PubAnnotationExportPipeline {
 		// we want to find documents that need BigQuery export
 		ProcessingStatusFlag targetProcessStatusFlag = ProcessingStatusFlag.BIGQUERY_LOAD_FILE_EXPORT_DONE;
 		// require that the documents be fully processed
-		Set<ProcessingStatusFlag> requiredProcessStatusFlags = EnumSet.of(ProcessingStatusFlag.TEXT_DONE,
-				ProcessingStatusFlag.OGER_CHEBI_DONE, ProcessingStatusFlag.OGER_CL_DONE,
-				ProcessingStatusFlag.OGER_GO_BP_DONE, ProcessingStatusFlag.OGER_GO_CC_DONE,
-				ProcessingStatusFlag.OGER_GO_MF_DONE, ProcessingStatusFlag.OGER_MOP_DONE,
-				ProcessingStatusFlag.OGER_NCBITAXON_DONE, ProcessingStatusFlag.OGER_PR_DONE,
-				ProcessingStatusFlag.OGER_SO_DONE, ProcessingStatusFlag.OGER_UBERON_DONE);
+		Set<ProcessingStatusFlag> requiredProcessStatusFlags = EnumSet.of(ProcessingStatusFlag.TEXT_DONE);
+
+//		,
+//				ProcessingStatusFlag.DP_DONE);
+//				,
+//				
+//				ProcessingStatusFlag.OGER_CHEBI_DONE, ProcessingStatusFlag.OGER_CL_DONE,
+//				ProcessingStatusFlag.OGER_GO_BP_DONE, ProcessingStatusFlag.OGER_GO_CC_DONE);
+//				ProcessingStatusFlag.OGER_GO_MF_DONE, ProcessingStatusFlag.OGER_MOP_DONE,
+//				ProcessingStatusFlag.OGER_NCBITAXON_DONE, ProcessingStatusFlag.OGER_PR_DONE,
+//				ProcessingStatusFlag.OGER_SO_DONE, ProcessingStatusFlag.OGER_UBERON_DONE);
 
 		// query Cloud Datastore to find document IDs in need of processing
 		DatastoreProcessingStatusUtil statusUtil = new DatastoreProcessingStatusUtil();
