+++ src/main/java/edu/cuanschutz/ccp/tm_provider/etl/deprecated/PubAnnotationExportPipeline.java
@@ -1,4 +1,4 @@
-package edu.cuanschutz.ccp.tm_provider.etl;
+package edu.cuanschutz.ccp.tm_provider.etl.deprecated;
 
 import java.util.ArrayList;
 import java.util.Arrays;
@@ -26,9 +26,12 @@ import org.apache.beam.sdk.values.PCollectionTuple;
 
 import com.google.datastore.v1.Entity;
 
+import edu.cuanschutz.ccp.tm_provider.etl.EtlFailureData;
+import edu.cuanschutz.ccp.tm_provider.etl.PipelineMain;
 import edu.cuanschutz.ccp.tm_provider.etl.fn.BigQueryExportFileBuilderFn;
 import edu.cuanschutz.ccp.tm_provider.etl.fn.DocumentDownloadFn;
 import edu.cuanschutz.ccp.tm_provider.etl.fn.EtlFailureToEntityFn;
+import edu.cuanschutz.ccp.tm_provider.etl.fn.PubAnnotationExportFileBuilderFn;
 import edu.cuanschutz.ccp.tm_provider.etl.util.DatastoreProcessingStatusUtil;
 import edu.cuanschutz.ccp.tm_provider.etl.util.DatastoreProcessingStatusUtil.OverwriteOutput;
 import edu.cuanschutz.ccp.tm_provider.etl.util.DocumentCriteria;
@@ -39,15 +42,15 @@ import edu.cuanschutz.ccp.tm_provider.etl.util.ProcessingStatusFlag;
 import edu.cuanschutz.ccp.tm_provider.etl.util.Version;
 
 /**
- * This Apache Beam pipeline processes documents with the OGER concept
- * recognition service reached via HTTP POST. Input is plain text; Output is
- * concept annotations in BioNLP format.
+ * Note: this pipeline is not actively used
+ * 
+ * This Apache Beam pipeline exports annotations in the PubAnnotation format.
  */
-public class BigQueryExportPipeline {
+public class PubAnnotationExportPipeline {
 
-	private final static Logger LOGGER = Logger.getLogger(BigQueryExportPipeline.class.getName());
+	private final static Logger LOGGER = Logger.getLogger(PubAnnotationExportPipeline.class.getName());
 
-	private static final PipelineKey PIPELINE_KEY = PipelineKey.BIGQUERY_EXPORT;
+	private static final PipelineKey PIPELINE_KEY = PipelineKey.PUBANNOTATION_EXPORT;
 
 	public interface Options extends DataflowPipelineOptions {
 		@Description("Location of the output bucket")
@@ -64,13 +67,14 @@ public class BigQueryExportPipeline {
 		OverwriteOutput getOverwrite();
 
 		void setOverwrite(OverwriteOutput value);
+
 	}
 
 	public static void main(String[] args) {
 		String pipelineVersion = Version.getProjectVersion();
 		com.google.cloud.Timestamp timestamp = com.google.cloud.Timestamp.now();
 		Options options = PipelineOptionsFactory.fromArgs(args).withValidation().as(Options.class);
-		LOGGER.log(Level.INFO, String.format("Running BigQuery export pipeline"));
+		LOGGER.log(Level.INFO, String.format("Running PubAnnontation export pipeline"));
 
 		Pipeline p = Pipeline.create(options);
 		PCollection<String> documentIds = getDocumentIdsToProcess(p, options.getCollection(), options.getOverwrite());
@@ -94,58 +98,37 @@ public class BigQueryExportPipeline {
 		nonredundantFailureEntities.apply("annot_extract_failure_entity->datastore",
 				DatastoreIO.v1().write().withProjectId(options.getProject()));
 
-		DocumentCriteria outputDocCriteria = new DocumentCriteria(DocumentType.BIGQUERY, DocumentFormat.BIGQUERY,
+		DocumentCriteria outputDocCriteria = new DocumentCriteria(DocumentType.PUBANNOTATION, DocumentFormat.JSON,
 				PIPELINE_KEY, pipelineVersion);
 
-		PCollectionTuple bigqueryExports = BigQueryExportFileBuilderFn.processNoTrackDocIds(docIdToAnnotations,
+		PCollectionTuple pubAnnotExports = PubAnnotationExportFileBuilderFn.processNoTrackDocIds(docIdToAnnotations,
 				outputDocCriteria, timestamp);
 
-		PCollection<String> docIdToBigQuery_annotationTable = bigqueryExports
-				.get(BigQueryExportFileBuilderFn.ANNOTATION_TABLE_OUTPUT_TAG_NO_TRACK_DOCID);
-		PCollection<String> docIdToBigQuery_inSectionTable = bigqueryExports
-				.get(BigQueryExportFileBuilderFn.IN_SECTION_TABLE_OUTPUT_TAG_NO_TRACK_DOCID);
-		PCollection<String> docIdToBigQuery_inParagraphTable = bigqueryExports
-				.get(BigQueryExportFileBuilderFn.IN_PARAGRAPH_TABLE_OUTPUT_TAG_NO_TRACK_DOCID);
-		PCollection<String> docIdToBigQuery_inSentenceTable = bigqueryExports
-				.get(BigQueryExportFileBuilderFn.IN_SENTENCE_TABLE_OUTPUT_TAG_NO_TRACK_DOCID);
-		PCollection<String> docIdToBigQuery_inConceptTable = bigqueryExports
-				.get(BigQueryExportFileBuilderFn.IN_CONCEPT_TABLE_OUTPUT_TAG_NO_TRACK_DOCID);
-		PCollection<String> docIdToBigQuery_relationTable = bigqueryExports
-				.get(BigQueryExportFileBuilderFn.RELATION_TABLE_OUTPUT_TAG_NO_TRACK_DOCID);
-		PCollection<EtlFailureData> bigqueryExportFailures = bigqueryExports
+		PCollection<String> docIdToJson = pubAnnotExports.get(PubAnnotationExportFileBuilderFn.JSON_OUTPUT_TAG);
+		PCollection<EtlFailureData> pubAnnotExportFailures = pubAnnotExports
 				.get(BigQueryExportFileBuilderFn.FAILURE_TAG);
 
 		/*
 		 * store the failures for this pipeline in Cloud Datastore - deduplication is
 		 * necessary to avoid Datastore non-transactional commit errors
 		 */
-		failureEntities = bigqueryExportFailures.apply("bq_export_failures->datastore",
+		failureEntities = pubAnnotExportFailures.apply("pubannot_export_failures->datastore",
 				ParDo.of(new EtlFailureToEntityFn()));
 		nonredundantFailureEntities = PipelineMain.deduplicateEntitiesByKey(failureEntities);
-		nonredundantFailureEntities.apply("bq_export_failure_entity->datastore",
+		nonredundantFailureEntities.apply("pubannot_export_failure_entity->datastore",
 				DatastoreIO.v1().write().withProjectId(options.getProject()));
 
 		// write the output
-		docIdToBigQuery_annotationTable.apply("write annotation table",
-				TextIO.write().to(options.getOutputBucket() + "/annotation.").withSuffix(".tsv"));
-		docIdToBigQuery_inSectionTable.apply("write annotation table",
-				TextIO.write().to(options.getOutputBucket() + "/in-section.").withSuffix(".tsv"));
-		docIdToBigQuery_inParagraphTable.apply("write annotation table",
-				TextIO.write().to(options.getOutputBucket() + "/in-paragraph.").withSuffix(".tsv"));
-		docIdToBigQuery_inSentenceTable.apply("write annotation table",
-				TextIO.write().to(options.getOutputBucket() + "/in-sentence.").withSuffix(".tsv"));
-		docIdToBigQuery_inConceptTable.apply("write annotation table",
-				TextIO.write().to(options.getOutputBucket() + "/in-concept.").withSuffix(".tsv"));
-		docIdToBigQuery_relationTable.apply("write annotation table",
-				TextIO.write().to(options.getOutputBucket() + "/relation.").withSuffix(".tsv"));
+		docIdToJson.apply("write pubannot json file",
+				TextIO.write().to(options.getOutputBucket() + "/pubannotation.").withSuffix(".json"));
 
 		PCollectionList<EtlFailureData> failureList = PCollectionList.of(annotationRetrievalFailures)
-				.and(bigqueryExportFailures);
+				.and(pubAnnotExportFailures);
 		PCollection<EtlFailureData> mergedFailures = failureList.apply(Flatten.<EtlFailureData>pCollections());
 
 		// update the status for documents that were successfully processed
 		PCollection<KV<String, String>> successStatus = DatastoreProcessingStatusUtil.getSuccessStatus(documentIds,
-				mergedFailures, ProcessingStatusFlag.BIGQUERY_LOAD_FILE_EXPORT_DONE);
+				mergedFailures, ProcessingStatusFlag.PUBANNOTATION_FILE_EXPORT_DONE);
 		List<PCollection<KV<String, String>>> statusList = new ArrayList<PCollection<KV<String, String>>>();
 		statusList.add(successStatus);
 		DatastoreProcessingStatusUtil.performStatusUpdatesInBatch(statusList);
@@ -157,12 +140,10 @@ public class BigQueryExportPipeline {
 	private static List<DocumentCriteria> populateDocumentCriteria(String pipelineVersion) {
 		List<DocumentCriteria> documentCriteria = Arrays.asList(
 				new DocumentCriteria(DocumentType.TEXT, DocumentFormat.TEXT, PipelineKey.BIOC_TO_TEXT, pipelineVersion),
-				new DocumentCriteria(DocumentType.SENTENCE, DocumentFormat.BIONLP, PipelineKey.SENTENCE_SEGMENTATION,
-						pipelineVersion),
-				new DocumentCriteria(DocumentType.SECTIONS, DocumentFormat.BIONLP, PipelineKey.BIOC_TO_TEXT,
-						pipelineVersion),
-				new DocumentCriteria(DocumentType.DEPENDENCY_PARSE, DocumentFormat.CONLLU, PipelineKey.DEPENDENCY_PARSE,
-						pipelineVersion),
+//				new DocumentCriteria(DocumentType.SECTIONS, DocumentFormat.BIONLP, PipelineKey.BIOC_TO_TEXT,
+//						pipelineVersion),
+//				new DocumentCriteria(DocumentType.DEPENDENCY_PARSE, DocumentFormat.CONLLU, PipelineKey.DEPENDENCY_PARSE,
+//						pipelineVersion),
 				new DocumentCriteria(DocumentType.CONCEPT_CHEBI, DocumentFormat.BIONLP, PipelineKey.OGER,
 						pipelineVersion),
 				new DocumentCriteria(DocumentType.CONCEPT_CL, DocumentFormat.BIONLP, PipelineKey.OGER, pipelineVersion),
@@ -192,17 +173,12 @@ public class BigQueryExportPipeline {
 		// we want to find documents that need BigQuery export
 		ProcessingStatusFlag targetProcessStatusFlag = ProcessingStatusFlag.BIGQUERY_LOAD_FILE_EXPORT_DONE;
 		// require that the documents be fully processed
-		Set<ProcessingStatusFlag> requiredProcessStatusFlags = EnumSet.of(ProcessingStatusFlag.TEXT_DONE);
-
-//		,
-//				ProcessingStatusFlag.DP_DONE);
-//				,
-//				
-//				ProcessingStatusFlag.OGER_CHEBI_DONE, ProcessingStatusFlag.OGER_CL_DONE,
-//				ProcessingStatusFlag.OGER_GO_BP_DONE, ProcessingStatusFlag.OGER_GO_CC_DONE);
-//				ProcessingStatusFlag.OGER_GO_MF_DONE, ProcessingStatusFlag.OGER_MOP_DONE,
-//				ProcessingStatusFlag.OGER_NCBITAXON_DONE, ProcessingStatusFlag.OGER_PR_DONE,
-//				ProcessingStatusFlag.OGER_SO_DONE, ProcessingStatusFlag.OGER_UBERON_DONE);
+		Set<ProcessingStatusFlag> requiredProcessStatusFlags = EnumSet.of(ProcessingStatusFlag.TEXT_DONE,
+				ProcessingStatusFlag.OGER_CHEBI_DONE, ProcessingStatusFlag.OGER_CL_DONE,
+				ProcessingStatusFlag.OGER_GO_BP_DONE, ProcessingStatusFlag.OGER_GO_CC_DONE,
+				ProcessingStatusFlag.OGER_GO_MF_DONE, ProcessingStatusFlag.OGER_MOP_DONE,
+				ProcessingStatusFlag.OGER_NCBITAXON_DONE, ProcessingStatusFlag.OGER_PR_DONE,
+				ProcessingStatusFlag.OGER_SO_DONE, ProcessingStatusFlag.OGER_UBERON_DONE);
 
 		// query Cloud Datastore to find document IDs in need of processing
 		DatastoreProcessingStatusUtil statusUtil = new DatastoreProcessingStatusUtil();
